+++ internal/checks/query_series.go
@@ -59,8 +59,7 @@ func (c SeriesCheck) Check(ctx context.Context, rule parser.Rule) (problems []Pr
 }
 
 func (c SeriesCheck) countSeries(ctx context.Context, expr parser.PromQLExpr, selector promParser.VectorSelector) (problems []Problem) {
-	q := fmt.Sprintf("count(%s)", selector.String())
-	qr, err := c.prom.Query(ctx, q)
+	sets, err := c.prom.Series(ctx, []string{selector.String()})
 	if err != nil {
 		problems = append(problems, Problem{
 			Fragment: selector.String(),
@@ -72,12 +71,7 @@ func (c SeriesCheck) countSeries(ctx context.Context, expr parser.PromQLExpr, se
 		return
 	}
 
-	var series int
-	for _, s := range qr.Series {
-		series += int(s.Value)
-	}
-
-	if series == 0 {
+	if len(sets) == 0 {
 		if len(selector.LabelMatchers) > 1 {
 			// retry selector with only __name__ label
 			s := stripLabels(selector)